+++ src/app/syllogimous/components/card/card.component.html
@@ -4,13 +4,13 @@
             <div>
                 <span
                     class="tier-badge badge clickable fs-6"
-                    [style.backgroundColor]="TIER_COLORS[sylSrv.tier].bgColor"
-                    [style.color]="TIER_COLORS[sylSrv.tier].textColor"
+                    [style.backgroundColor]="TIER_COLORS[game.tier].bgColor"
+                    [style.color]="TIER_COLORS[game.tier].textColor"
                     [ngbTooltip]="tierTooltip"
                     [autoClose]="'outside'"
                     triggers="click"
                 >
-                    {{ sylSrv.tier }} <i class="bi bi-info-circle"></i>
+                    {{ game.tier }} <i class="bi bi-info-circle"></i>
                 </span>
                 <ng-template #tierTooltip>
                     <div class="text-start">
@@ -19,7 +19,7 @@
                         </div>
                     </div>
                 </ng-template>
-                <span class="text-muted ms-2">{{ sylSrv.score }} pts</span>
+                <span class="text-muted ms-2">{{ game.score }} pts</span>
             </div>
             <app-card-dropdown></app-card-dropdown>
         </div>
