+++ src/app/syllogimous/components/progress-and-performance/progress-and-performance.component.html
@@ -75,7 +75,7 @@
                     <span class="px-2">
                         {{trainingUnitPercentage.wrong}}
                     </span>
-                    <span class="progress-label">{{ sylSrv.question.type }}</span>
+                    <span class="progress-label">{{ game.question.type }}</span>
                 </div>
             </div>
         </ng-container>
