+++ src/app/syllogimous/constants/game.constants.ts
