+++ src/app/syllogimous/pages/history/history.component.html
@@ -2,7 +2,7 @@
     <div body class="p-3">
         <div class="text-center" *ngIf="!questions.length">
             <h3 class="mb-3">Play to see something here</h3>
-            <button class="btn btn-lg btn-primary" (click)="sylSrv.playArcadeMode()">
+            <button class="btn btn-lg btn-primary" (click)="game.playArcadeMode()">
                 Play <i class="bi bi-play-fill"></i>
             </button>
         </div>
