+++ src/app/syllogimous/pages/intro/intro.component.html
@@ -24,7 +24,7 @@
         <div>
             <button
                 class="btn btn-outline-primary"
-                (click)="sylSrv.skipIntro(dontShowAnymore.checked)"
+                (click)="game.skipIntro(dontShowAnymore.checked)"
             >Skip Intro</button>
         </div>
     </div>
