+++ src/app/syllogimous/pages/settings/settings.component.html
@@ -96,7 +96,7 @@
         <div>
             <button
                 class="btn btn-primary"
-                (click)="sylSrv.playArcadeMode()"
+                (click)="game.playArcadeMode()"
             >
                 Play <i class="bi bi-play-fill"></i>
             </button>
