+++ src/app/syllogimous/pages/stats/tier-stats/tier-stats.component.html
@@ -3,12 +3,12 @@
 <ul class="list-group">
     <li class="list-group-item flex-between">
         <b>Tier</b>
-        <div>{{ sylSrv.tier }}</div>
+        <div>{{ game.tier }}</div>
     </li>
 
     <li class="list-group-item flex-between">
         <b>Points Earned</b>
-        <div>{{ sylSrv.score }}</div>
+        <div>{{ game.score }}</div>
     </li>
 
     <li class="list-group-item flex-between">
