+++ src/app/syllogimous/pages/stats/type-based-stats/type-based-stats.component.html
@@ -1,7 +1,7 @@
 <div class="card p-3 rounded text-center" *ngIf="!types.length">
     <h2>Type-Based</h2>
     <h4 class="my-3">Play to see something here</h4>
-    <button class="btn btn-lg btn-primary" (click)="sylSrv.playArcadeMode()">
+    <button class="btn btn-lg btn-primary" (click)="game.playArcadeMode()">
         Play <i class="bi bi-play-fill"></i>
     </button>
 </div>
