+++ src/app/syllogimous/pages/tutorial/tutorial.component.html
@@ -15,7 +15,7 @@
         <div>
             <button
                 class="btn btn-outline-primary"
-                (click)="sylSrv.skipTutorial(dontShowAnymore.checked)"
+                (click)="game.skipTutorial(dontShowAnymore.checked)"
             >Skip Tutorial</button>
         </div>
     </div>
