+++ src/app/syllogimous/pages/tutorial/tutorial.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
-import { SyllogimousService } from '../../services/syllogimous.service';
+import { GameService } from '../../services/game.service';
 import { Router } from '@angular/router';
-import { EnumScreens } from '../../constants/syllogimous.constants';
+import { EnumScreens } from '../../constants/game.constants';
 
 @Component({
     selector: 'app-tutorial',
@@ -14,7 +14,7 @@ export class TutorialComponent {
     showBack = false;
 
     constructor(
-        public sylSrv: SyllogimousService,
+        public game: GameService,
         public router: Router
     ) {
         const navigation = this.router.getCurrentNavigation();
