+++ src/app/syllogimous/pages/tutorials/tutorials.component.html
@@ -2,7 +2,7 @@
     <div body class="py-3 px-md-3">
         <div class="text-center" *ngIf="!questions.length">
             <h3 class="mb-3">Play to discover question types</h3>
-            <button class="btn btn-lg btn-primary" (click)="sylSrv.playArcadeMode()">
+            <button class="btn btn-lg btn-primary" (click)="game.playArcadeMode()">
                 Play <i class="bi bi-play-fill"></i>
             </button>
         </div>
